implemented history sorting and minor bug fix

+++ domain/src/main/java/rahulstech/android/expensetracker/domain/impl/AccountRepositoryImpl.kt
@@ -73,7 +73,7 @@ internal class AccountRepositoryImpl(
         val account = accountDao.findAccountById(id)
         account?.let {
             val account = it.toAccount()
-            val updatedAccount = account.copy(balance = account.balance.toFloat() + amount.toFloat())
+            val updatedAccount = account.copy(balance = account.balance + amount.toFloat())
             updateAccount(updatedAccount)
         }
     }
